pyproject.toml
[tool.poetry]
name = "tutorial-kafka-python"
version = "0.1.0"
description = ""
authors = ["MasatoShima <masato.shima@quick.jp>"]

[tool.poetry.dependencies]
python = "^3.8.0"
kafka-python = "^2.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
